Add web worker entrypoint for WasmDatabase
This commit introduces a new Dart file `worker.dart` that serves as the entrypoint for a web worker. The worker will be compiled to JavaScript and is designed to facilitate the use of the WasmDatabase in a Flutter web application. The main function calls `WasmDatabase.workerMainForOpen()` to initialize the worker.